requirements.txt
Django==4.0.6
boto3==1.24.33
django-storages==1.12.3
django-environ==0.9.0
djangorestframework==3.13.1